web_app/requirements.txt
# Web框架
Flask==3.0.0
Flask-SQLAlchemy==3.1.1

# 用户认证
Flask-Login==0.6.3
bcrypt==4.1.2

# 数据库
pymysql==1.1.0
SQLAlchemy==2.0.23

# 配置管理
python-dotenv==1.0.0

# CORS支持
Flask-CORS==4.0.0

# 模板引擎（Flask自带Jinja2）
Jinja2==3.1.2

# 表单处理
WTForms==3.1.1
Flask-WTF==1.2.1
